requirements.txt
mysql-connector-python==8.0.22
PyYAML==5.3.1